refactor ConferenceSection
